project wide rename from mescal to mesqual
