+++ docs/index.md
@@ -1,3 +1,3 @@
-# Welcome to the MESCAL Vanilla Documentation
-This documentation provides a detailed overview of the mescal ecosystem, 
+# Welcome to the MESQUAL Vanilla Documentation
+This documentation provides a detailed overview of the mesqual ecosystem, 
 including tutorials, examples, and a complete API reference.
