+++ studies/mkdocs_study_01.yml
@@ -1,6 +1,6 @@
-site_name: 'mescal study_01'
+site_name: 'mesqual study_01'
 
-docs_dir: ./study_01_intro_to_mescal
+docs_dir: ./study_01_intro_to_mesqual
 
 theme:
   name: material
@@ -26,13 +26,13 @@ plugins:
         - "*.py"
 
 nav:
-  - Notebook 101: notebooks/mescal_101_study_manager_and_basic_fetching.ipynb
-  - Notebook 102: notebooks/mescal_102_mastering_data_fetching.ipynb
-  - Notebook 103: notebooks/mescal_103_scenario_attributes.ipynb
-  - Notebook 301: notebooks/mescal_301_time_series_dashboard.ipynb
-  - Notebook 302: notebooks/mescal_302_segmented_colormap.ipynb
-  - Notebook 303: notebooks/mescal_303_folium_model_df_map.ipynb
-  - Notebook 304: notebooks/mescal_304_folium_area_kpi_map.ipynb
-  - Notebook 304: notebooks/mescal_304_simple_area_kpi_map.ipynb
-  - Notebook 307: notebooks/mescal_307_country_plotter_util.ipynb
-  - Notebook 308: notebooks/mescal_308_html_dashboards.ipynb
+  - Notebook 101: notebooks/mesqual_101_study_manager_and_basic_fetching.ipynb
+  - Notebook 102: notebooks/mesqual_102_mastering_data_fetching.ipynb
+  - Notebook 103: notebooks/mesqual_103_scenario_attributes.ipynb
+  - Notebook 301: notebooks/mesqual_301_time_series_dashboard.ipynb
+  - Notebook 302: notebooks/mesqual_302_segmented_colormap.ipynb
+  - Notebook 303: notebooks/mesqual_303_folium_model_df_map.ipynb
+  - Notebook 304: notebooks/mesqual_304_folium_area_kpi_map.ipynb
+  - Notebook 304: notebooks/mesqual_304_simple_area_kpi_map.ipynb
+  - Notebook 307: notebooks/mesqual_307_country_plotter_util.ipynb
+  - Notebook 308: notebooks/mesqual_308_html_dashboards.ipynb
