+++ studies/study_01_intro_to_mesqual/data/networks_acdc_example/cheap_wind.nc
