+++ studies/study_01_intro_to_mesqual/data/networks_scigrid_de/solar_150.nc
