+++ studies/study_01_intro_to_mesqual/data/networks_scigrid_de/wind_150.nc
